+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2026-03-04
+
 ### Changed
 
 - `#[tool]` macro now derives `Serialize`, `Clone`, and `Debug` for generated ToolArgs structs